refact: SceneDelegate에 있던 탭바 설정 코드를 다른 파일로 분리
